Modularise support state
Fixes #42484.
