+++ client/state/support/constants.js
@@ -1,0 +1,3 @@
+export const SESSION_NONE = 'none';
+export const SESSION_ACTIVE = 'active';
+export const SESSION_EXPIRED = 'expired';
