+++ client/state/support/selectors.js
@@ -1,7 +1,9 @@
 /**
  * Internal dependencies
  */
-import { SESSION_ACTIVE, SESSION_EXPIRED } from './reducer';
+import { SESSION_ACTIVE, SESSION_EXPIRED } from './constants';
+
+import 'calypso/state/support/init';
 
 export function isSupportSession( { support } ) {
 	return support === SESSION_ACTIVE || support === SESSION_EXPIRED;
